TATTVA_SEMICOLON_2024/tattva/requirements.txt
black==24.2.0
blinker==1.7.0
boto3==1.34.49
botocore==1.34.49
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dash==2.15.0
dash-bootstrap-components==0.13.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.2
fonttools==4.49.0
idna==3.6
igraph==0.11.4
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
jaal==0.1.5
Jinja2==3.1.3
jmespath==1.0.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
plotly==5.19.0
PyJWT==2.8.0
pyparsing==3.1.1
python-arango==7.9.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
requests-toolbelt==1.0.0
retrying==1.3.4
s3transfer==0.10.0
six==1.16.0
tenacity==8.2.3
texttable==1.7.0
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
visdcc==0.0.50
Werkzeug==3.0.1
zipp==3.17.0